Week4/requirements.txt
easyocr==1.7.1
editdistance==0.6.2
matplotlib==3.7.3
numpy==1.24.4
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
optuna==3.3.0
pandas==2.0.3
Pillow==10.0.1
Pillow==10.1.0
pytesseract==0.3.10
python_Levenshtein==0.23.0
scipy==1.11.3
skimage==0.0
tqdm==4.66.1
Levenshtein==0.23.0
